we are SO BACK
= change: changed some menu storage

= change: removed type conversions of cutNoise and converted type to float

= todo: cry

+++ src/main/kotlin/utils/storage/globals.kt
@@ -26,4 +26,4 @@ var squareRows: LockType<Int> = LockType(15)
 var squareColumns: LockType<Int> = LockType(15)
 
 // Cut Noise
-var cutNoise: LockType<Double> = LockType(0.0)
+var cutNoise: LockType<Float> = LockType(0.0F)
